refactor(upgrade): split migration logic into multiple files
